fix: put the road on top of its bridge, and keep the deck level
Two defects behind one symptom: a span whose carriageway ran underneath
its own structure, bowed across the water.

Order. The road surface sampler reads deck heights out of the client
grid mirror, but onSnapshot meshed the road before applying the road
deltas to that mirror. At mesh time every bridge tile still read
elevation 0, so the carriageway baked at terrain height — on the
riverbed — while the structure, lamps and furniture, all built after the
mirror updated, stood at deck height. The mirror now goes first.

applyRoadDeltas returns the tiles whose height moved, which both forces
that order and gives the caller what it cannot otherwise know: tiles
beside a deck sample the ramp blend but carry no delta of their own, so
their geometry only rebuilds if asked.

Precision. Every consumer reconstructs a deck as terrain + offset, and
the offset was stored a byte per tile. A solved-flat 5.00 m deck came
back as 4.65..5.47 — the rounding error tracks the riverbed's fraction,
bowing the span by up to half a metre. The layer widens to Float32 and
the solver stores the exact lift, with anything under a centimetre of it
still counting as at grade. SAVE_VERSION 4 -> 5; a v4 save's whole-metre
bytes widen into the new layer, so saved bridges keep their height.

The regression test reconstructs the deck the way a renderer does and
asserts a dished bed comes out level; it fails if the rounding returns.

+++ src/sim/dispatch.test.ts
@@ -74,7 +74,7 @@ function makeGrid(): GridState {
     fields: Array.from({ length: FIELD_COUNT }, () => new Uint8Array(n)),
     district: new Uint8Array(n),
     landfill: new Uint8Array(n),
-    roadElevation: new Uint8Array(n),
+    roadElevation: new Float32Array(n),
   };
 }
